Add `local-server` subcommand to `remote_settings` script (#212)
* Update Python Dependencies

Adds the following dependencies:

* logging
* pydantic
* pyyaml

* Move create subcommand to its own directory

Moves the `create` subcommand, which used to just be a single file,
to its own directory. I think the structure will be cleaner if each
subcommand has its own directory even if it only has the one file.

* Add local-server subcommand to remote_settings

Adds a new subcommand to the RemoteSettings CLI tool to spin
up the local server. This is basically a pared-down copy of
the local remote settings script from the translations repo.

* Rename test_create.py

Renames test_create to reflect that it is testing a mocked server.

* Add test_create_local_server.py

Adds a new test file for testing the local remote settings server.
This has all the same assertions as the mocked version. I may remove
the mocked version in the future as the project evolves, but for now
I'm just keeping bothof them 1:1 as is.

* Conslidate test constants

Conslidates test constants that are shared between both
the mocked tests and the local server tests.

* Enable local-server tests in CI

These are all the changes that were required to get
the local-server tests to build and run in CI.

* Skip evaluate if models directory is not modified

Ensures that the evaluation step will not run unless
the `models` directory was modified. This project will
see more changes in the coming months, and I don't want
to waste CPU cycles for no-op evals.

* Load schemas and display fields on local server

Ensures that the JSON schemas and display fields are loaded
into the new collections when the local server is created.

This makes the UI look much nicer when opening the admin
page in the browser.

* Remove logging dependency

* Expand comments and READMEs

* Update heartbeat comment